Move UnitGroupController and UnitGroups into namespaced folder
